+++ RRLab/R/libraryR.R
@@ -12,6 +12,9 @@
 #' # Also works with unquoted names (non-standard evaluation)
 #' libraryR(dplyr)
 #'
+#' # Calling it on RRLab itself triggers a colourful message
+#' libraryR("RRLab")
+#'
 #' # Package names can also be stored in a vector
 #' pkgs <- c("ggplot2", "dplyr")
 #' libraryR(pkgs)
@@ -26,7 +29,7 @@ libraryR = function(pkg) {
   }
   
   pkg = substitute(pkg)
-  pkg = if (is.name(pkg)) as.character(pkg) else eval(pkg)
+  pkg = if (is.name(pkg)) as.character(pkg) else eval(pkg, parent.frame())
   
   results = setNames(logical(length(pkg)), pkg)
   
@@ -60,6 +63,23 @@ libraryR = function(pkg) {
     
     suppressPackageStartupMessages(library(p, character.only = TRUE))
     cli::cli_alert_success("Loaded {.pkg {p}}{if (installed_now) ' after install' else ''}.")
+
+    if (identical(p, "RRLab")) {
+      msg <- "~ Achievement unlocked: Super user of libraryR()! ~"
+      color_set <- c(32, 33, 35, 36, 31, 34)
+      n <- nchar(msg)
+
+      for (iter in rep(seq_along(color_set),3)) {
+        cat("\r")
+        cat(paste0("\033[", color_set[iter], "m", msg, "\033[0m"))
+        
+        Sys.sleep(0.1)
+        
+        flush.console()
+    }
+      cat("\n")
+    }
+
     results[p] = TRUE
   }