+++ NEWS.md
@@ -1,7 +1,8 @@
 # RRLab Release Notes
 
 ## 0.0.0.1001
-- Check for RRLab updates on package attach using the installed `GithubSHA1`.
+- Check for newer RRLab versions on package attach.
+
 
 ## 0.0.0.1000
 - Initial public release of RRLab.